+++ src/main/java/common/dao/IUserInfoDAO.java
@@ -3,5 +3,5 @@ package common.dao;
 import java.util.List;
 
 public interface IUserInfoDAO {
-    List<Integer> getAllUserID();
+    List<Long> getAllUserID();
 }
