+++ src/main/java/yankunwei/dao/UserInfoDAO.java
@@ -16,8 +16,8 @@ public class UserInfoDAO implements IUserInfoDAO {
     private Logger logger = LogManager.getLogger(UserInfoDAO.class);
     
     @Override
-    public List<Integer> getAllUserID() {
-        List<Integer> userIDs = new ArrayList<>();
+    public List<Long> getAllUserID() {
+        List<Long> userIDs = new ArrayList<>();
         Connection connection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
@@ -28,7 +28,7 @@ public class UserInfoDAO implements IUserInfoDAO {
             preparedStatement = connection.prepareStatement(SQL);
             resultSet = preparedStatement.executeQuery();
             while (resultSet.next()) {
-                userIDs.add(resultSet.getInt("UserID"));
+                userIDs.add(resultSet.getLong("UserID"));
             }
         } catch (SQLException e) {
             logger.error("Query All UserID failed");
